Re-structured files correctly
